requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
pymongo==4.0.1
python-dotenv==0.19.1
gunicorn==20.1.0
requests==2.26.0
dnspython==2.1.0 